Refactor MuTE processing logic
