+++ src/main/java/gregtech/api/multitileentity/multiblock/base/Controller.java
@@ -65,8 +65,8 @@ import gregtech.api.logic.ControllerFluidLogic;
 import gregtech.api.logic.ControllerItemLogic;
 import gregtech.api.logic.FluidInventoryLogic;
 import gregtech.api.logic.ItemInventoryLogic;
+import gregtech.api.logic.MuTEProcessingLogic;
 import gregtech.api.logic.PowerLogic;
-import gregtech.api.logic.ProcessingLogic;
 import gregtech.api.logic.interfaces.PowerLogicHost;
 import gregtech.api.multitileentity.enums.MultiTileCasingPurpose;
 import gregtech.api.multitileentity.interfaces.IMultiBlockController;
@@ -87,7 +87,7 @@ import mcp.mobius.waila.api.IWailaDataAccessor;
 /**
  * Multi Tile Entities - or MuTEs - don't have dedicated hatches, but their casings can become hatches.
  */
-public abstract class Controller<T extends Controller<T, P>, P extends ProcessingLogic<P>>
+public abstract class Controller<T extends Controller<T, P>, P extends MuTEProcessingLogic<P>>
     extends MultiTileBasicMachine<P> implements IAlignment, IMultiBlockController, IDescribable, IMTE_AddToolTips,
     ISurvivalConstructable, ControllerWithOptionalFeatures, IGlobalWirelessEnergy {
