+++ src/main/java/gregtech/api/multitileentity/multiblock/base/StackableModularController.java
@@ -6,11 +6,11 @@ import java.util.Map;
 
 import org.jetbrains.annotations.NotNull;
 
-import gregtech.api.logic.ProcessingLogic;
+import gregtech.api.logic.MuTEProcessingLogic;
 import gregtech.api.multitileentity.interfaces.UpgradableModularMuTE;
 import gregtech.api.util.GT_StructureUtilityMuTE.UpgradeCasings;
 
-public abstract class StackableModularController<T extends StackableModularController<T, P>, P extends ProcessingLogic<P>>
+public abstract class StackableModularController<T extends StackableModularController<T, P>, P extends MuTEProcessingLogic<P>>
     extends StackableController<T, P> implements UpgradableModularMuTE {
 
     protected double durationMultiplier = 1;
