+++ src/main/java/gregtech/common/tileentities/machines/multiblock/logic/MaceratorProcessingLogic.java
@@ -1,7 +1,7 @@
 package gregtech.common.tileentities.machines.multiblock.logic;
 
-import gregtech.api.logic.ProcessingLogic;
+import gregtech.api.logic.MuTEProcessingLogic;
 
-public class MaceratorProcessingLogic extends ProcessingLogic<MaceratorProcessingLogic> {
+public class MaceratorProcessingLogic extends MuTEProcessingLogic<MaceratorProcessingLogic> {
 
 }
